Backend/bidding/requirements.txt
APScheduler==3.10.4
asgiref==3.7.2
Django==5.0
django-apscheduler==0.6.2
django-cors-headers==4.3.1
django-redis==5.4.0
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
install==1.3.5
Pillow==10.1.0
psycopg==3.1.14
psycopg2==2.9.9
PyJWT==1.7.1
pytz==2023.3.post1
redis==5.0.1
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
tzlocal==5.2
